Move TestDefaultBinaryMessenger to its own file
